+++ AirLib/src/vehicles/multirotor/api/MultirotorRpcLibClient.cpp
@@ -121,7 +121,7 @@ MultirotorRpcLibClient* MultirotorRpcLibClient::moveToPositionAsync(float x, flo
     return this;
 }
 
-MultirotorRpcLibClient* MultirotorRpcLibClient::moveToGlobalPositionAsync(float lat, float lon, float height, float velocity, float timeout_sec, 
+MultirotorRpcLibClient* MultirotorRpcLibClient::moveToGlobalPositionAsync(double lat, double lon, float height, float velocity, float timeout_sec, 
     DrivetrainType drivetrain, const YawMode& yaw_mode, float lookahead, float adaptive_lookahead, const std::string& vehicle_name)
 {
     pimpl_->last_future = static_cast<rpc::client*>(getClient())->async_call("moveToGlobalPosition", lat, lon, height, velocity, timeout_sec,
